refactor: Edit unit test code to check method 'GetAllMoves'
Due to the modifications that I made, member variables of the class 'Move' cannot be accessed directly. 

So I changed few lines of code to ensure it working properly.  

Also, my previous commit includes the addition of member variables as well as corresponding getters, and setters for them. So I added lines of code to test them.