refactor: Edit unit test code for 'GetAllMoves'
After building the project and running the executable 'UnitTest', I realized that there are some moves without accuracy stats (i.e. 'Guardian of Alola' one of the Z-Move introduced in Generation VII).

New standard for testing the code seems to be required. And I decided not to change it until I fix. (to alert myself)